+++ CHANGELOG.md
@@ -6,7 +6,7 @@ Notable changes will be documented in this file. The project adheres to [Semanti
 Unreleased
 ----------
 
-* Added `Page` and `PageLanguageOverlay` model
+* Added `Page` and `LanguageOverlay` model
 * Added common files
 * First commit